choir: remove unnecessary imports
