+++ src-tauri/src/airdrop.rs
@@ -1,8 +1,6 @@
 use jsonwebtoken::{decode, Algorithm, DecodingKey, Validation};
-use log::{info, warn};
+use log::warn;
 use serde::{Deserialize, Serialize};
-
-use crate::{mm_proxy_adapter::MergeMiningProxyConfig, UniverseAppState};
 
 const LOG_TARGET: &str = "tari::universe::airdrop";
 
@@ -66,30 +64,4 @@ pub async fn validate_jwt(airdrop_access_token: Option<String>) -> Option<String
             None
         }
     })
-}
-
-pub async fn restart_mm_proxy_with_new_telemetry_id(
-    state: tauri::State<'_, UniverseAppState>,
-) -> Result<(), String> {
-    let telemetry_id = state
-        .telemetry_manager
-        .read()
-        .await
-        .get_unique_string()
-        .await;
-    info!(target: LOG_TARGET, "getting new telemetry id -after {:?}", telemetry_id);
-    let mm_proxy_manager_config = state
-        .mm_proxy_manager
-        .config()
-        .await
-        .ok_or("mm proxy config could not be found")?;
-    let _ = state
-        .mm_proxy_manager
-        .change_config(MergeMiningProxyConfig {
-            coinbase_extra: telemetry_id.clone(),
-            ..mm_proxy_manager_config
-        })
-        .await
-        .map_err(|e| e.to_string());
-    Ok(())
 }
