+++ src-tauri/src/commands_internal.rs
@@ -4,7 +4,9 @@ use log::{error, info, warn};
 use tauri::Manager;
 use tokio::time::Instant;
 
-use crate::{gpu_miner_adapter::GpuNodeSource, UniverseAppState};
+use crate::{
+    gpu_miner_adapter::GpuNodeSource, mm_proxy_adapter::MergeMiningProxyConfig, UniverseAppState,
+};
 
 const LOG_TARGET: &str = "tari::universe::mining_operations";
 const MAX_ACCEPTABLE_COMMAND_TIME: Duration = Duration::from_secs(1);
@@ -166,5 +168,31 @@ pub async fn start_mining<'r>(
     if timer.elapsed() > MAX_ACCEPTABLE_COMMAND_TIME {
         warn!(target: LOG_TARGET, "start_mining took too long: {:?}", timer.elapsed());
     }
+    Ok(())
+}
+
+pub async fn restart_mm_proxy_with_new_telemetry_id(
+    state: tauri::State<'_, UniverseAppState>,
+) -> Result<(), String> {
+    let telemetry_id = state
+        .telemetry_manager
+        .read()
+        .await
+        .get_unique_string()
+        .await;
+    info!(target: LOG_TARGET, "getting new telemetry id -after {:?}", telemetry_id);
+    let mm_proxy_manager_config = state
+        .mm_proxy_manager
+        .config()
+        .await
+        .ok_or("mm proxy config could not be found")?;
+    let _ = state
+        .mm_proxy_manager
+        .change_config(MergeMiningProxyConfig {
+            coinbase_extra: telemetry_id.clone(),
+            ..mm_proxy_manager_config
+        })
+        .await
+        .map_err(|e| e.to_string());
     Ok(())
 }
