+++ src-tauri/src/main.rs
@@ -51,7 +51,7 @@ use tari_common::configuration::Network;
 use tari_common_types::tari_address::TariAddress;
 use tari_shutdown::Shutdown;
 use tauri::async_runtime::{block_on, JoinHandle};
-use tauri::{Emitter, Listener, Manager, RunEvent};
+use tauri::{Emitter, Manager, RunEvent};
 use tauri_plugin_sentry::{minidump, sentry};
 use tokio::select;
 use tokio::sync::{Mutex, RwLock};
@@ -92,6 +92,7 @@ mod app_in_memory_config;
 mod auto_launcher;
 mod binaries;
 mod commands;
+mod commands_internal;
 mod consts;
 mod cpu_miner;
 mod credential_manager;
@@ -104,7 +105,6 @@ mod gpu_miner;
 mod gpu_miner_adapter;
 mod hardware;
 mod internal_wallet;
-mod mining_operations;
 mod mm_proxy_adapter;
 mod mm_proxy_manager;
 mod network_utils;
